+++ sketch_sep16a/Zuurstofsensor.h
@@ -9,7 +9,7 @@ private:
 		Zuurstofsensor(int);
 		~Zuurstofsensor();
 		
-    int16_t Meet(float temperatuur);
+    int Meet(float temperatuur);
 };
 
 #endif
